✨ Feat: border-bottom 추가로 경계 생성
